postgres tweaks

+++ server/models/shape.py
@@ -7,8 +7,8 @@ from server import Base
 class Shape(Base):
   __tablename__ = 'shapes'
   shape_id = Column(String(50), primary_key=True)
-  shape_pt_lat = Column(Float(precision=64))
-  shape_pt_lon = Column(Float(precision=64))
+  shape_pt_lat = Column(Float(precision=53))
+  shape_pt_lon = Column(Float(precision=53))
   shape_pt_time = Column(String(50))
   shape_pt_sequence = Column(Integer, primary_key=True)
